Import Book8088 BIOS v1.0.2 12/20/22 (MM/DD/YY)
This came from a ZIP file, `111book8088 hand386.zip` sent to someone
who purchased a Book8088. Within it it contained an archive
`BOOK8088Temporary Technician Profile.rar`, which in turn contained a
`BOOK8088技术员资料临时/B8088-BIOS1.0` directory with this source code.

This has been imported on top of commit 1dc3c95 dated 2022-12-14, as
that's the last commit in Sergey's repo before the 2022-12-20 date
that this BIOS has in config.inc. (The next commit Sergey's repo is
2023-01-05.)